requirements.txt
requests==2.24.0
pymongo==3.11.0
python-dotenv==0.14.0
discord.py==1.4.1